# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TRC Control Plane API
  version: "v1"
  description: |
    Control plane HTTP API.

    Key properties:
    - Org-scoped paths for tenant resources.
    - Append-only event log as source of truth, materialized views for reads.
    - Workloads run as Firecracker microVM instances on hosts.
    - Ingress is L4-first (TLS SNI passthrough by default).
servers:
  - url: /v1
...
security:
  - bearerAuth: []
paths:
...
  /orgs/{org_id}/apps/{app_id}/envs:
    get:
      tags: [Envs]
      summary: List environments
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Envs
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvsResponse"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
    post:
      tags: [Envs]
      summary: Create environment
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnvRequest"
      responses:
        "200":
          description: Env created
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Env"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "409":
          $ref: "#/components/responses/Error409"
  /orgs/{org_id}/apps/{app_id}/envs/{env_id}:
    get:
      tags: [Envs]
      summary: Get environment
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/EnvId"
      responses:
        "200":
          description: Env
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Env"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
    patch:
      tags: [Envs]
      summary: Update environment
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/EnvId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnvRequest"
      responses:
        "200":
          description: Env updated
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Env"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
    delete:
      tags: [Envs]
      summary: Delete environment
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/EnvId"
      responses:
        "200":
          description: Deleted (idempotent)
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  headers:
    XRequestId:
      description: Request correlation id
      schema:
        type: string
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Opaque idempotency key for retry-safe writes
      schema:
        type: string
        minLength: 8
        maxLength: 128
    OrgId:
      name: org_id
      in: path
      required: true
      schema:
        type: string
...
    AppId:
      name: app_id
      in: path
      required: true
      schema:
        type: string
    EnvId:
      name: env_id
      in: path
      required: true
      schema:
        type: string
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
    DeleteResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    Env:
      type: object
      required: [id, app_id, name, created_at]
      properties:
        id:
          type: string
        app_id:
          type: string
        name:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
        resource_version:
          type: integer
    ListEnvsResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Env"
        next_cursor:
          type: [string, "null"]
    CreateEnvRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateEnvRequest:
      type: object
      required: [expected_version]
      properties:
        name:
          type: string
        expected_version:
          type: integer
          minimum: 0
...